fix(ui): resolve remaining text layout review findings
